added portfolio functionality, still need to add stock to portfolio

+++ backend/db/migrations/20201124164649-create-portfolio.js
@@ -14,7 +14,7 @@ module.exports = {
         references: {model: "Users"}
       },
       buyingPower: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.NUMERIC,
         defaultValue: 2000 
       },
       createdAt: {
